Implement calcDisplay to display numbers
The calcDisplay div will be updated whenever the input is updated,
an operation is calculated, or the calculator is cleared/reset. It
would be nice to implement an operator display as well, either in the
display window or by a button press. There is also a bug where selecting
multiple operators after a value is calculated. The second operation
calculates using an undefined second operand.